bot/requirements.txt
firebase_admin>=6.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
fake-useragent>=1.4.0
